perevalapi/requirements.txt
asgiref==3.8.1
django==5.2.2
django-filter==25.1
djangorestframework==3.16.0
markdown==3.8
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
